Split up trigger connection into two files
